refactor benchmark file structure
